+++ CHANGELOG.md
@@ -16,6 +16,8 @@ But it is not used in between anymore. This is resolving a lot of inconsistencie
 > You should not experience any breakage if you are merely using the questionnaire filler components.
 > You will see numerous API changes which will require fixes if you have forked/modified.
 
+* WIP: repeating question items: Repeating items can be removed now. This is buggy and leads to inconsistent data model!
+
 ## 0.5.4-dev.6
 * Fix bug with unit extension
 * WIP: repeating question items